fix: harden fuzzer-found integer-overflow + getMutable defects (vm/analysis/compiler)
A batch of fixes from the fuzzer plus a static audit of the same defect classes.

Integer overflow/underflow — panic with overflow-checks, UB / silent-wrap in C++
(the C++ ones filed as upstream Issues 5-6):
- bit32 negation of INT_MIN (lshift/rshift/arshift/rrotate): `unsigned_abs()` /
  `wrapping_neg()`
- bit32 extract/replace `f + w` overflow — runtime (fieldargs/luau_f_extract/
  luau_f_replace) AND the compiler constant-fold (compile_expr_call): widen to i64
- os.time `year - 1900` / `month - 1` on INT_MIN: `wrapping_sub`
- int64.extract/replace `f + w`: compare `w > 64 - f`
- stacklimitreached `stack_last - top` usize underflow: compute signed (isize)

getMutable-needs-follow class (LUAU_ASSERT abort under armed assertions):
- PromoteTypeLevels TypeId TableType + FunctionType arms: add the `is::<BoundType>`
  short-circuit (siblings of the already-fixed FreeType / FreeTypePack arms)
- delete 6 dead duplicate inherent-method files that re-embodied the same bug

Other port deviations from C++:
- unifier_try_unify_variadics: null/ICE check moved before the deref (C++ order)
- magic_format (old + new solver): guard `args.size == 0` underflow/OOB on
  zero-arg self-calls

Regression tests added under `--features typecheck`. Full suite: 5656 tests pass;
all overflow repros clean and valid bit32/os.time values verified correct.

+++ crates/luaur-compiler/src/methods/compiler_compile_expr_call.rs
@@ -145,7 +145,9 @@ impl Compiler {
                 } else {
                     -1
                 };
-                if fi >= 0 && wi > 0 && fi + wi <= 32 {
+                // Widen the add: `fi`/`wi` are folded user constants and `fi + wi`
+                // overflows `int` for huge fields (UB in C++; panic w/ overflow-checks).
+                if fi >= 0 && wi > 0 && fi as i64 + wi as i64 <= 32 {
                     let fwp = fi | ((wi - 1) << 5);
                     let cid = (*self.bytecode).add_constant_number(fwp as f64);
                     if cid >= 0 {
